+++ src/main/java/io/xlorey/FluxLoader/annotations/SubscribeEvent.java
@@ -11,5 +11,5 @@ public @interface SubscribeEvent {
      * Event name
      * @return The default event name for the subscription. If not specified, an empty string is returned.
      */
-    String eventName() default "";
+    String eventName();
 }