+++ src/main/java/io/xlorey/FluxLoader/utils/BackupTools.java
@@ -1,7 +1,9 @@
 package io.xlorey.FluxLoader.utils;
 
 import lombok.experimental.UtilityClass;
+import org.objectweb.asm.*;
 
+import java.io.FileInputStream;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -39,7 +41,7 @@ public class BackupTools {
                 }
             });
         } catch (IOException e) {
-            Logger.print("Error while listing files in the directory.");
+            Logger.printSystem("Error while listing files in the directory.");
             throw e;
         }
         return additionalFiles;
@@ -55,7 +57,7 @@ public class BackupTools {
         for (Path file : additionalFiles) {
             Path backupPath = file.resolveSibling(file.getFileName() + BACKUP_EXTENSION);
             if (Files.exists(backupPath)) {
-                Logger.print(String.format("Backup of the file '%s' already exists. Skipping backup...", file.getFileName()));
+                Logger.printSystem(String.format("Backup of the file '%s' already exists. Skipping backup...", file.getFileName()));
             } else {
                 validatedFiles.add(file);
             }
@@ -73,7 +75,7 @@ public class BackupTools {
         for (Path file : additionalFiles) {
             Path backupPath = file.resolveSibling(file.getFileName() + BACKUP_EXTENSION);
             if (!Files.exists(backupPath)) {
-                Logger.print(String.format("Backup file for '%s' not found. Skipping restore...", file.getFileName()));
+                Logger.printSystem(String.format("Backup file for '%s' not found. Skipping restore...", file.getFileName()));
             } else {
                 validatedFiles.add(file);
             }
@@ -82,16 +84,49 @@ public class BackupTools {
     }
 
     /**
+     * Checks for the presence of the Modified annotation in the class file's methods.
+     * This method reads a class file specified by the given path and analyzes it to
+     * determine if any of its methods are annotated with the Modified annotation.
+     * If such an annotation is found, it logs a warning and throws an IllegalStateException.
+     * @param classFilePath The path to the class file to be checked.
+     * @throws IOException If an I/O error occurs during reading the class file.
+     */
+    private static void checkAnnotationModified(Path classFilePath) throws IOException {
+        try (FileInputStream fileInputStream = new FileInputStream(classFilePath.toFile())) {
+            ClassReader reader = new ClassReader(fileInputStream);
+
+            reader.accept(new ClassVisitor(Opcodes.ASM9) {
+                @Override
+                public MethodVisitor visitMethod(int access, String name, String descriptor, String signature, String[] exceptions) {
+                    return new MethodVisitor(Opcodes.ASM9) {
+                        @Override
+                        public AnnotationVisitor visitAnnotation(String descriptor, boolean visible) {
+                            if (descriptor.equals("Lio/xlorey/FluxLoader/annotations/Modified;")) {
+                                Logger.printSystem("Signs of modification detected! Please, before installing a new version, remove the old one using the '--uninstall' flag!");
+                                throw new IllegalStateException(String.format("Annotation 'Modified' found in method '%s' of file '%s'", name, classFilePath.toAbsolutePath()));
+                            }
+                            return super.visitAnnotation(descriptor, visible);
+                        }
+                    };
+                }
+            }, 0);
+        }
+    }
+
+    /**
      * Creating a backup file
      * @param pathToClassFile path to the file with '.class' extension
      * @throws IOException in cases of unsuccessful backup creation
      */
     public static void createBackup(String pathToClassFile) throws IOException {
-        Logger.print(String.format("Trying to create a restore point for a file '%s'...", pathToClassFile));
+        Logger.printSystem(String.format("Trying to create a restore point for a file '%s'...", pathToClassFile));
 
         validateClassFilePath(pathToClassFile);
 
         Path originalFilePath = getFullPath(pathToClassFile);
+
+        checkAnnotationModified(originalFilePath);
+
         Path backupFilePath = getBackupPath(originalFilePath);
 
         Path directoryPath = originalFilePath.getParent();
@@ -99,24 +134,24 @@ public class BackupTools {
         ArrayList<Path> additionalFiles = validateAdditionalFilesForBackup(getAdditionalFiles(directoryPath, originalFilePath));
 
         if (Files.exists(backupFilePath)) {
-            Logger.print(String.format("Backup of the file '%s' already exists. Skipping backup...", originalFilePath.getFileName()));
+            Logger.printSystem(String.format("Backup of the file '%s' already exists. Skipping backup...", originalFilePath.getFileName()));
             return;
         }
 
         try {
             Files.copy(originalFilePath, backupFilePath);
 
-            Logger.print(String.format("The backup for the main file '%s' has been successfully created!", originalFilePath.getFileName()));
+            Logger.printSystem(String.format("The backup for the main file '%s' has been successfully created!", originalFilePath.getFileName()));
 
             for (Path file : additionalFiles) {
                 Path backupPath = file.resolveSibling(file.getFileName() + BACKUP_EXTENSION);
                 Files.copy(file, backupPath);
 
-                Logger.print(String.format("Backup for dependent file '%s' was successfully created!", file.getFileName()));
+                Logger.printSystem(String.format("Backup for dependent file '%s' was successfully created!", file.getFileName()));
             }
 
         } catch (IOException e) {
-            Logger.print("Error while creating backup file.");
+            Logger.printSystem("Error while creating backup file.");
             throw e;
         }
 
@@ -128,7 +163,7 @@ public class BackupTools {
      * @throws IOException in cases of unsuccessful recovery
      */
     public static void restoreFile(String pathToClassFile) throws IOException {
-        Logger.print(String.format("Attempting to recover a file '%s'...", pathToClassFile));
+        Logger.printSystem(String.format("Attempting to recover a file '%s'...", pathToClassFile));
 
         validateClassFilePath(pathToClassFile);
 
@@ -140,23 +175,23 @@ public class BackupTools {
         ArrayList<Path> additionalFiles = validateAdditionalFilesForRestore(getAdditionalFiles(directoryPath, originalFilePath));
 
         if (!Files.exists(backupFilePath)) {
-            Logger.print(String.format("Backup file for '%s' not found. Skipping restore...", originalFilePath.getFileName()));
+            Logger.printSystem(String.format("Backup file for '%s' not found. Skipping restore...", originalFilePath.getFileName()));
             return;
         }
 
         try {
             Files.move(backupFilePath, originalFilePath, StandardCopyOption.REPLACE_EXISTING);
 
-            Logger.print(String.format("The main file '%s' was successfully restored!", originalFilePath.getFileName()));
+            Logger.printSystem(String.format("The main file '%s' was successfully restored!", originalFilePath.getFileName()));
 
             for (Path file : additionalFiles) {
                 Path backupPath = file.resolveSibling(file.getFileName() + BACKUP_EXTENSION);
                 Files.move(backupPath, file, StandardCopyOption.REPLACE_EXISTING);
 
-                Logger.print(String.format("Dependent file '%s' was successfully restored!", file.getFileName()));
+                Logger.printSystem(String.format("Dependent file '%s' was successfully restored!", file.getFileName()));
             }
         } catch (IOException e) {
-            Logger.print("Error when restoring the game file.");
+            Logger.printSystem("Error when restoring the game file.");
             throw e;
         }